added damage increase for skills based on attributes and skill effects

+++ src/main/java/de/raidcraft/skills/api/ability/Ability.java
@@ -55,7 +55,7 @@ public interface Ability<T extends CharacterTemplate> {
 
     boolean isOfElement(EffectElement element);
 
-    int getTotalDamage();
+    double getTotalDamage();
 
     double getTotalCastTime();
